manual labeling from kaggle general faces dataset
this notebook is made because we felt like it is quite time consuming if we move the images data one by one just using a conventional way like moving the images from one folder to a filtered folder in a local file manager. This notebook also highlight bright and dark spot on each face images so it can help others to define which type of skin the image is. also there is a button which will automatically copy the image to its label in separate folder